Removed unused code for geolocation. Now is required a credit card number to use geocode API, this functionality must be updated.
